pyproject.toml
[tool.poetry]
name = "chat-app-plus"
version = "0.1.0"
description = ""
authors = ["gorbart <bartek.go3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
channels = "^4.0.0"
yapf = "^0.32.0"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"
daphne = "^4.0.0"
redis = "^4.4.2"
channels-redis = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
